requirements.txt
numpy>=1.24.0
fastapi>=0.115.0
uvicorn>=0.34.0
mistralai>=1.0.0
python-dotenv>=1.0.0
yfinance>=0.2.36
